Move methods from array_schema to experimental
