+++ tiledb/sm/cpp_api/tiledb_experimental
@@ -45,5 +45,7 @@
 #include "query_experimental.h"
 #include "subarray_experimental.h"
 #include "vfs_experimental.h"
+#include "ndrectangle.h"
+#include "current_domain.h"
 
 #endif  // TILEDB_EXPERIMENTAL_CPP_H
